[#8] refactor : parse 도메인 생성, 하위 input&output 패키지로 나눠서 패키지 구조 리팩토링
