+++ src/main/java/org/unilab/improfessorbe/domain/parse/input/convert/PdfFileParser.java
@@ -1,4 +1,4 @@
-package org.unilab.improfessorbe.domain.problem.infrastructure.external.parser;
+package org.unilab.improfessorbe.domain.parse.input.convert;
 
 import java.io.IOException;
